feat: move converter to own folder
