+++ src/Contracts/Convertible.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Innobrain\Structure\Contracts;
 
+use Innobrain\Structure\Converters\Concerns\ConvertStrategy;
+
 /**
  * Marks any object that can be converted using a ConvertStrategy.
  */
