+++ src/Converters/Array/ArrayPermittedValue.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Innobrain\Structure\Converters\Array;
+
+use Innobrain\Structure\DTOs\PermittedValue;
+
+trait ArrayPermittedValue
+{
+    public function convertPermittedValue(PermittedValue $permittedValue): array
+    {
+        return $this->normalize([
+            'key' => $permittedValue->key,
+            'label' => $permittedValue->label,
+        ]);
+    }
+}
