+++ src/Converters/JsonSchema/JsonSchemaConvertStrategy.php
@@ -1,0 +1,36 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Innobrain\Structure\Converters\JsonSchema;
+
+use Innobrain\Structure\Converters\Concerns\ConvertStrategy;
+
+/**
+ * Convert the package's DTOs into Prism PHP schemas.
+ *
+ * Typical usage:
+ *   $schema = $module->convert(new JsonSchemaConvertStrategy());
+ *   $fieldSchema = $field->convert(new JsonSchemaConvertStrategy());
+ *
+ * The strategy returns:
+ *   • Module   ⇒ ObjectType with properties for each field
+ *   • Field    ⇒ Type (type depends on field type)
+ */
+final readonly class JsonSchemaConvertStrategy implements ConvertStrategy
+{
+    use JsonSchemaField;
+    use JsonSchemaFieldDependency;
+    use JsonSchemaFieldFilter;
+    use JsonSchemaModule;
+    use JsonSchemaPermittedValue;
+
+    /**
+     * @param  bool  $includeNullable  true ➜ mark fields as nullable when they have no default
+     * @param  bool  $includeDescriptions  true ➜ include field labels as descriptions
+     */
+    public function __construct(
+        private bool $includeNullable = true, // @phpstan-ignore-line
+        private bool $includeDescriptions = true, // @phpstan-ignore-line
+    ) {}
+}
