+++ src/Converters/LaravelRules/LaravelRulesFieldDependency.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Innobrain\Structure\Converters\LaravelRules;
+
+use Innobrain\Structure\DTOs\FieldDependency;
+
+trait LaravelRulesFieldDependency
+{
+    public function convertFieldDependency(FieldDependency $fieldDependency): array
+    {
+        return [];
+    }
+}
