+++ src/Converters/LaravelRules/LaravelRulesFieldFilter.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Innobrain\Structure\Converters\LaravelRules;
+
+use Innobrain\Structure\DTOs\FieldFilter;
+
+trait LaravelRulesFieldFilter
+{
+    public function convertFieldFilter(FieldFilter $fieldFilter): array
+    {
+        return [];
+    }
+}
