+++ src/Converters/LaravelRules/LaravelRulesModule.php
@@ -1,0 +1,39 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Innobrain\Structure\Converters\LaravelRules;
+
+use Innobrain\Structure\DTOs\Field;
+use Innobrain\Structure\DTOs\Module;
+use Innobrain\Structure\Enums\FieldType;
+
+trait LaravelRulesModule
+{
+    /**
+     * The returned array is keyed by field key. For multi-select fields an
+     * additional "{fieldKey}.*" element is added so each item receives the
+     * proper "in:" check.
+     *
+     * @return array<string, string|array>
+     */
+    public function convertModule(Module $module): array
+    {
+        $result = [];
+
+        foreach ($module->fields as $fieldKey => $field) {
+            /** @var Field $field */
+            $result[$fieldKey] = $this->convertField($field);
+
+            // Extra rules per item for multi-selects
+            if ($field->type === FieldType::MultiSelect && $field->permittedValues->isNotEmpty()) {
+                $itemRules = [
+                    'in:'.implode(',', $field->permittedValues->keys()->toArray()),
+                ];
+                $result[$fieldKey.'.*'] = $this->pipeOrArray($itemRules);
+            }
+        }
+
+        return $result;
+    }
+}
