+++ src/Enums/FieldConfigurationModule.php
@@ -17,4 +17,11 @@ enum FieldConfigurationModule: string
     case Project = 'project';
     case Task = 'task';
     case User = 'user';
+
+    public static function values(array $only = []): array
+    {
+        $all = array_map(static fn (self $element) => $element->value, self::cases());
+
+        return $only === [] ? $all : array_values(array_intersect($all, $only));
+    }
 }
