+++ src/Services/FieldConfiguration.php
@@ -7,6 +7,7 @@ namespace Innobrain\Structure\Services;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 use Innobrain\OnOfficeAdapter\Dtos\OnOfficeApiCredentials;
+use Innobrain\OnOfficeAdapter\Exceptions\OnOfficeException;
 use Innobrain\OnOfficeAdapter\Facades\FieldRepository;
 use Innobrain\Structure\Collections\FieldCollection;
 use Innobrain\Structure\Collections\ModulesCollection;
@@ -26,17 +27,12 @@ class FieldConfiguration
 {
     /**
      * Retrieve the field configuration for a given client.
+     *
+     * @throws OnOfficeException
      */
     public function retrieveForClient(OnOfficeApiCredentials $credentials, array $only = []): ModulesCollection
     {
-        $moduleCases = FieldConfigurationModule::cases();
-        $moduleValues = collect($moduleCases)
-            ->map(fn (FieldConfigurationModule $module) => $module->value)
-            ->toArray();
-
-        if (count($only) > 0) {
-            $moduleValues = array_filter($moduleValues, fn (string $value) => in_array($value, $only, true));
-        }
+        $moduleValues = FieldConfigurationModule::values($only);
 
         $rawModulesData = FieldRepository::query()
             ->withCredentials($credentials)
@@ -59,7 +55,6 @@ class FieldConfiguration
             if (! is_array($moduleData['elements'])) {
                 continue;
             }
-
             $moduleKey = $moduleData['id'] ?? $moduleKey;
             $moduleEnum = FieldConfigurationModule::tryFrom((string) $moduleKey);
 
@@ -100,17 +95,23 @@ class FieldConfiguration
                 continue; // unknown field type
             }
 
-            $fields->put((string) $fieldKey, new Field(
-                key: (string) $fieldKey,
-                label: (string) Arr::get($fieldData, 'label', ucfirst((string) $fieldKey)),
+            $fields->put($fieldKey, new Field(
+                key: $fieldKey,
+                label: (string) Arr::get($fieldData, 'label', ucfirst($fieldKey)),
                 type: $fieldType,
-                length: Arr::get($fieldData, 'length') ? (int) Arr::get($fieldData, 'length') : null,
+                length: Arr::get($fieldData, 'length')
+                    ? (int) Arr::get($fieldData, 'length')
+                    : null,
                 permittedValues: $this->parsePermittedValues(Arr::get($fieldData, 'permittedvalues', [])),
-                default: Arr::get($fieldData, 'default') ? (string) Arr::get($fieldData, 'default') : null,
+                default: Arr::get($fieldData, 'default')
+                    ? (string) Arr::get($fieldData, 'default')
+                    : null,
                 filters: $this->parseFieldFilters(Arr::get($fieldData, 'filters', [])),
                 dependencies: $this->parseFieldDependencies(Arr::get($fieldData, 'dependencies', [])),
                 compoundFields: collect(Arr::get($fieldData, 'compoundFields', [])),
-                fieldMeasureFormat: Arr::get($fieldData, 'fieldMeasureFormat') ? (string) Arr::get($fieldData, 'fieldMeasureFormat') : null,
+                fieldMeasureFormat: Arr::get($fieldData, 'fieldMeasureFormat')
+                    ? (string) Arr::get($fieldData, 'fieldMeasureFormat')
+                    : null,
             ));
         }
