+++ src/StructureServiceProvider.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Innobrain\Structure;
 
-use Innobrain\Structure\Commands\StructureCommand;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
