+++ tests/Feature/PrismSchemaConvertStrategyFeatureTest.php
@@ -2,10 +2,9 @@
 
 declare(strict_types=1);
 
-use Illuminate\Support\Collection;
 use Innobrain\Structure\Collections\FieldCollection;
 use Innobrain\Structure\Collections\ModulesCollection;
-use Innobrain\Structure\Converters\PrismSchemaConvertStrategy;
+use Innobrain\Structure\Converters\PrismSchema\PrismSchemaConvertStrategy;
 use Innobrain\Structure\DTOs\Field;
 use Innobrain\Structure\DTOs\FieldDependency;
 use Innobrain\Structure\DTOs\Module;
